refactor: convert to workspace; extract pre-encoders into ore-encoders crate
Reshape the repo as a Cargo workspace under `packages/`:

- `packages/ore-rs/` — the existing ORE library, unchanged in API or
  behaviour. Carries its own README and CHANGELOG.
- `packages/ore-encoders/` — new crate hosting canonical
  order-preserving pre-encoders for plaintext types fed into the ORE
  machinery. Each encoder is gated behind a per-type feature so callers
  only pay for what they use.

The `Decimal` pre-encoder (`pre_encode`, the constants, the
constant-time `strip_trailing_zeros` / `digit_count` helpers, and the
structural / equivalence / byte-order tests) moves into
`ore-encoders::decimal`. The `OreEncrypt for Decimal` impl and all
ORE-level tests (quickcheck properties, ciphertext ordering, hex
round-trip) stay in `ore-rs`, which now depends on `ore-encoders`
behind its `decimal` feature. The split avoids a circular dep — the
encoder is pure plaintext-to-bytes, with no ORE types in scope — and
gives future encoders an obvious home.

Workspace root holds only the `[workspace]` table and the
`bumpalo` patch (also workspace-scoped). License file stays at the
root and is referenced by both crates via relative path.

Test counts split as expected — `ore-encoders` runs the 7 pre-encoder
tests, `ore-rs` runs the 56 ORE-level unit tests + 5 doctests; total
63, unchanged from before the split.
