+++ Cargo.toml
@@ -1,50 +1,10 @@
-[package]
-name = "ore-rs"
-version = "0.8.0"
-authors = ["Dan Draper <dan@cipherstash.com>"]
-edition = "2018"
-homepage = "https://cipherstash.com"
-description = "Order-revealing encryption library used by the CipherStash searchable encryption platform"
-license-file = "LICENCE"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[dev-dependencies]
-criterion = "0.5.1"
-quickcheck = "1.0.3"
-# Test-only: macros for constructing `Decimal` literals in unit tests.
-rust_decimal_macros = "1.28"
-
-[features]
-# Adds ORE support for `chrono::NaiveDate` and `chrono::DateTime<Utc>`.
-chrono = ["dep:chrono"]
-# Adds ORE support for `rust_decimal::Decimal` via a fixed-length 14-byte
-# order-preserving pre-encoding fed into the existing fixed-N ORE machinery.
-decimal = ["dep:rust_decimal"]
-
-[dependencies]
-aes = { version = "0.8.2", features = ["zeroize"] }
-block-modes = "0.9.1"
-byteorder = "1.4.3"
-chrono = { version = "0.4.42", default-features = false, optional = true }
-hex-literal = "0.4.1"
-rand = "0.8.5"
-rand_chacha = "0.3.1"
-rust_decimal = { version = "1.28", default-features = false, optional = true }
-num = "0.4.0"
-hex = "0.4.3"
-subtle-ng = "2.5.0"
-zeroize = { version = "1.5.7", features = ["zeroize_derive"] }
-lazy_static = "1.4.0"
-thiserror = "1.0.38"
+[workspace]
+resolver = "2"
+members = [
+    "packages/ore-rs",
+    "packages/ore-encoders",
+]
 
 [patch.crates-io]
 # version 3.16.0
 bumpalo = { git = 'https://github.com/fitzgen/bumpalo.git', rev = "4eeab88" }
-
-
-[[bench]]
-name = "oreaes128"
-harness = false
-
-[[example]]
-name = "encrypt"
