+++ packages/ore-rs/src/primitives/prp.rs
