+++ packages/ore-rs/src/scheme/bit2.rs
